Finishing Feature Filter Calender in History
