--- app/src/main/java/com/arkan/terbangin/presentation/calenderfilterhistory/Example4Fragment.kt
@@ -1,286 +1,0 @@
-package com.arkan.terbangin.presentation.calenderfilterhistory
-
-import android.graphics.Color
-import android.graphics.drawable.Drawable
-import android.os.Build
-import android.os.Bundle
-import android.util.TypedValue
-import android.view.Menu
-import android.view.MenuInflater
-import android.view.View
-import android.widget.TextView
-import androidx.annotation.RequiresApi
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.graphics.BlendModeColorFilterCompat
-import androidx.core.graphics.BlendModeCompat
-import androidx.core.view.children
-import com.arkan.terbangin.R
-import com.arkan.terbangin.databinding.Example4CalendarDayBinding
-import com.arkan.terbangin.databinding.Example4CalendarHeaderBinding
-import com.arkan.terbangin.databinding.Example4FragmentBinding
-import com.arkan.terbangin.presentation.calenderfilterhistory.shared.ContinuousSelectionHelper.getSelection
-import com.arkan.terbangin.presentation.calenderfilterhistory.shared.ContinuousSelectionHelper.isInDateBetweenSelection
-import com.arkan.terbangin.presentation.calenderfilterhistory.shared.ContinuousSelectionHelper.isOutDateBetweenSelection
-import com.arkan.terbangin.presentation.calenderfilterhistory.shared.DateSelection
-import com.arkan.terbangin.presentation.calenderfilterhistory.shared.dateRangeDisplayText
-import com.arkan.terbangin.presentation.calenderfilterhistory.shared.displayText
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.google.android.material.snackbar.Snackbar
-import com.kizitonwose.calendar.core.CalendarDay
-import com.kizitonwose.calendar.core.CalendarMonth
-import com.kizitonwose.calendar.core.DayPosition
-import com.kizitonwose.calendar.core.daysOfWeek
-import com.kizitonwose.calendar.view.MonthDayBinder
-import com.kizitonwose.calendar.view.MonthHeaderFooterBinder
-import com.kizitonwose.calendar.view.ViewContainer
-import java.time.LocalDate
-import java.time.YearMonth
-import java.time.format.DateTimeFormatter
-
-class Example4Fragment : BaseFragment(R.layout.example_4_fragment), HasBackButton {
-    override val titleRes: Int? = null
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private val today = LocalDate.now()
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private var selection = DateSelection()
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private val headerDateFormatter = DateTimeFormatter.ofPattern("EEE'\n'd MMM")
-
-    private lateinit var binding: Example4FragmentBinding
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    override fun onViewCreated(
-        view: View,
-        savedInstanceState: Bundle?,
-    ) {
-        super.onViewCreated(view, savedInstanceState)
-        addStatusBarColorUpdate(R.color.white)
-        setHasOptionsMenu(true)
-        binding = Example4FragmentBinding.bind(view)
-        // Set the First day of week depending on Locale
-        val daysOfWeek = daysOfWeek()
-        binding.legendLayout.root.children.forEachIndexed { index, child ->
-            (child as TextView).apply {
-                text = daysOfWeek[index].displayText()
-                setTextSize(TypedValue.COMPLEX_UNIT_SP, 15f)
-                setTextColorRes(R.color.md_theme_outline)
-            }
-        }
-
-        configureBinders()
-        val currentMonth = YearMonth.now()
-        binding.exFourCalendar.setup(
-            currentMonth,
-            currentMonth.plusMonths(12),
-            daysOfWeek.first(),
-        )
-        binding.exFourCalendar.scrollToMonth(currentMonth)
-
-        binding.exFourSaveButton.setOnClickListener click@{
-            val (startDate, endDate) = selection
-            if (startDate != null && endDate != null) {
-                val text = dateRangeDisplayText(startDate, endDate)
-                Snackbar.make(requireView(), text, Snackbar.LENGTH_LONG).show()
-            }
-            parentFragmentManager.popBackStack()
-        }
-
-        bindSummaryViews()
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun bindSummaryViews() {
-        binding.exFourStartDateText.apply {
-            if (selection.startDate != null) {
-                text = headerDateFormatter.format(selection.startDate)
-                setTextColorRes(R.color.md_theme_outline)
-            } else {
-                text = getString(R.string.text_start_date)
-                setTextColor(Color.GRAY)
-            }
-        }
-
-        binding.exFourEndDateText.apply {
-            if (selection.endDate != null) {
-                text = headerDateFormatter.format(selection.endDate)
-                setTextColorRes(R.color.md_theme_outline)
-            } else {
-                text = getString(R.string.text_end_date)
-                setTextColor(Color.GRAY)
-            }
-        }
-
-        binding.exFourSaveButton.isEnabled = selection.daysBetween != null
-    }
-
-    @Deprecated("Deprecated in Java")
-    @RequiresApi(Build.VERSION_CODES.O)
-    override fun onCreateOptionsMenu(
-        menu: Menu,
-        inflater: MenuInflater,
-    ) {
-        inflater.inflate(R.menu.example_4_menu, menu)
-        binding.exFourToolbar.post {
-            // Configure menu text to match what is in the Airbnb app.
-            binding.exFourToolbar.findViewById<TextView>(R.id.menuItemClear).apply {
-                setTextColor(requireContext().getColorCompat(R.color.md_theme_outline))
-                setTextSize(TypedValue.COMPLEX_UNIT_SP, 16f)
-                isAllCaps = false
-            }
-        }
-        menu.findItem(R.id.menuItemClear).setOnMenuItemClickListener {
-            selection = DateSelection()
-            binding.exFourCalendar.notifyCalendarChanged()
-            bindSummaryViews()
-            true
-        }
-    }
-
-    override fun onStart() {
-        super.onStart()
-        val closeIndicator =
-            requireContext().getDrawableCompat(R.drawable.ic_clear).apply {
-                colorFilter =
-                    BlendModeColorFilterCompat.createBlendModeColorFilterCompat(
-                        requireContext().getColorCompat(R.color.md_theme_outline),
-                        BlendModeCompat.SRC_ATOP,
-                    )
-            }
-        (activity as AppCompatActivity).supportActionBar?.setHomeAsUpIndicator(closeIndicator)
-    }
-
-    private fun configureBinders() {
-        val clipLevelHalf = 5000
-        val ctx = requireContext()
-        val rangeStartBackground =
-            ctx.getDrawableCompat(R.drawable.example_4_continuous_selected_bg_start).also {
-                it.level = clipLevelHalf // Used by ClipDrawable
-            }
-        val rangeEndBackground =
-            ctx.getDrawableCompat(R.drawable.example_4_continuous_selected_bg_end).also {
-                it.level = clipLevelHalf // Used by ClipDrawable
-            }
-        val rangeMiddleBackground =
-            ctx.getDrawableCompat(R.drawable.example_4_continuous_selected_bg_middle)
-        val singleBackground = ctx.getDrawableCompat(R.drawable.example_4_single_selected_bg)
-        val todayBackground = ctx.getDrawableCompat(R.drawable.example_4_today_bg)
-
-        @RequiresApi(Build.VERSION_CODES.O)
-        class DayViewContainer(view: View) : ViewContainer(view) {
-            lateinit var day: CalendarDay // Will be set when this container is bound.
-            val binding = Example4CalendarDayBinding.bind(view)
-
-            init {
-                view.setOnClickListener {
-                    if (day.position == DayPosition.MonthDate &&
-                        (day.date == today || day.date.isAfter(today))
-                    ) {
-                        selection =
-                            getSelection(
-                                clickedDate = day.date,
-                                dateSelection = selection,
-                            )
-                        this@Example4Fragment.binding.exFourCalendar.notifyCalendarChanged()
-                        bindSummaryViews()
-                    }
-                }
-            }
-        }
-
-        binding.exFourCalendar.dayBinder =
-            object : MonthDayBinder<DayViewContainer> {
-                @RequiresApi(Build.VERSION_CODES.O)
-                override fun create(view: View) = DayViewContainer(view)
-
-                @RequiresApi(Build.VERSION_CODES.O)
-                override fun bind(
-                    container: DayViewContainer,
-                    data: CalendarDay,
-                ) {
-                    container.day = data
-                    val textView = container.binding.exFourDayText
-                    val roundBgView = container.binding.exFourRoundBackgroundView
-                    val continuousBgView = container.binding.exFourContinuousBackgroundView
-
-                    textView.text = null
-                    roundBgView.makeInVisible()
-                    continuousBgView.makeInVisible()
-
-                    val (startDate, endDate) = selection
-
-                    when (data.position) {
-                        DayPosition.MonthDate -> {
-                            textView.text = data.date.dayOfMonth.toString()
-                            if (data.date.isBefore(today)) {
-                                textView.setTextColorRes(R.color.md_theme_outlineVariant)
-                            } else {
-                                when {
-                                    startDate == data.date && endDate == null -> {
-                                        textView.setTextColorRes(R.color.white)
-                                        roundBgView.applyBackground(singleBackground)
-                                    }
-                                    data.date == startDate -> {
-                                        textView.setTextColorRes(R.color.white)
-                                        continuousBgView.applyBackground(rangeStartBackground)
-                                        roundBgView.applyBackground(singleBackground)
-                                    }
-                                    startDate != null && endDate != null && (data.date > startDate && data.date < endDate) -> {
-                                        textView.setTextColorRes(R.color.md_theme_outline)
-                                        continuousBgView.applyBackground(rangeMiddleBackground)
-                                    }
-                                    data.date == endDate -> {
-                                        textView.setTextColorRes(R.color.white)
-                                        continuousBgView.applyBackground(rangeEndBackground)
-                                        roundBgView.applyBackground(singleBackground)
-                                    }
-                                    data.date == today -> {
-                                        textView.setTextColorRes(R.color.md_theme_outline)
-                                        roundBgView.applyBackground(todayBackground)
-                                    }
-                                    else -> textView.setTextColorRes(R.color.md_theme_outline)
-                                }
-                            }
-                        }
-                        // Make the coloured selection background continuous on the
-                        // invisible in and out dates across various months.
-                        DayPosition.InDate ->
-                            if (startDate != null && endDate != null &&
-                                isInDateBetweenSelection(data.date, startDate, endDate)
-                            ) {
-                                continuousBgView.applyBackground(rangeMiddleBackground)
-                            }
-                        DayPosition.OutDate ->
-                            if (startDate != null && endDate != null &&
-                                isOutDateBetweenSelection(data.date, startDate, endDate)
-                            ) {
-                                continuousBgView.applyBackground(rangeMiddleBackground)
-                            }
-                    }
-                }
-
-                private fun View.applyBackground(drawable: Drawable) {
-                    makeVisible()
-                    background = drawable
-                }
-            }
-
-        class MonthViewContainer(view: View) : ViewContainer(view) {
-            val textView = Example4CalendarHeaderBinding.bind(view).exFourHeaderText
-        }
-        binding.exFourCalendar.monthHeaderBinder =
-            object : MonthHeaderFooterBinder<MonthViewContainer> {
-                override fun create(view: View) = MonthViewContainer(view)
-
-                @RequiresApi(Build.VERSION_CODES.O)
-                override fun bind(
-                    container: MonthViewContainer,
-                    data: CalendarMonth,
-                ) {
-                    container.textView.text = data.yearMonth.displayText()
-                }
-            }
-    }
-}
