+++ app/src/main/java/com/arkan/terbangin/presentation/calenderfilterhistory/shared/Utils.kt
@@ -21,13 +21,13 @@ fun YearMonth.displayText(short: Boolean = false): String {
 @RequiresApi(Build.VERSION_CODES.O)
 fun Month.displayText(short: Boolean = true): String {
     val style = if (short) TextStyle.SHORT else TextStyle.FULL
-    return getDisplayName(style, Locale.ENGLISH)
+    return getDisplayName(style, Locale.getDefault())
 }
 
 @RequiresApi(Build.VERSION_CODES.O)
 fun DayOfWeek.displayText(uppercase: Boolean = false): String {
-    return getDisplayName(TextStyle.SHORT, Locale.ENGLISH).let { value ->
-        if (uppercase) value.uppercase(Locale.ENGLISH) else value
+    return getDisplayName(TextStyle.SHORT, Locale.getDefault()).let { value ->
+        if (uppercase) value.uppercase(Locale.getDefault()) else value
     }
 }
