+++ app/src/main/res/drawable/calendar_continuous_selected_bg_middle.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@color/md_theme_scrim" />
+    <solid android:color="@color/md_theme_inversePrimary" />
 </shape>
