+++ app/src/main/res/drawable/calendar_continuous_selected_bg_start.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <clip xmlns:android="http://schemas.android.com/apk/res/android"
     android:clipOrientation="horizontal"
-    android:gravity="start">
+    android:gravity="end">
     <shape android:shape="rectangle">
-        <solid android:color="@color/md_theme_scrim" />
+        <solid android:color="@color/md_theme_inversePrimary" />
     </shape>
 </clip>
