+++ app/src/main/res/drawable/calendar_today_bg.xml
@@ -3,5 +3,5 @@
     android:shape="oval">
     <stroke
         android:width="1dp"
-        android:color="@color/md_theme_outlineVariant" />
+        android:color="@color/md_theme_onBackground" />
 </shape>
