+++ app/src/main/res/layout/bottom_sheet_calender_filter_history.xml
@@ -1,13 +1,151 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:padding="16dp"
+    android:layout_height="match_parent"
+    android:background="@color/md_theme_background"
+    android:clickable="true"
+    android:focusable="true"
+    android:orientation="vertical"
     tools:context=".presentation.calenderfilterhistory.CalenderFilterHistoryBottomSheet">
 
+    <com.google.android.material.appbar.AppBarLayout
+        android:id="@+id/exFourAppBarLayout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/md_theme_background"
+        android:theme="@style/AppTheme.AppBarOverlay">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:paddingStart="14dp"
+            android:paddingTop="2dp"
+            android:paddingEnd="14dp"
+            android:paddingBottom="10dp">
+
+            <ImageView
+                android:id="@+id/iv_close"
+                android:layout_width="52dp"
+                android:layout_height="52dp"
+                android:layout_marginTop="8dp"
+                android:padding="12dp"
+                android:src="@drawable/ic_clear"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:tint="@color/md_theme_onSurfaceVariant" />
+
+            <TextView
+                android:id="@+id/tv_start_from"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/text_mulai_dari"
+                android:textColor="@color/md_theme_onBackground"
+                android:textSize="16sp"
+                android:layout_marginBottom="8dp"
+                app:layout_constraintStart_toStartOf="@id/exFourStartDateText"
+                app:layout_constraintEnd_toEndOf="@id/exFourStartDateText"
+                app:layout_constraintBottom_toTopOf="@id/exFourStartDateText"/>
+
+            <TextView
+                android:id="@+id/exFourStartDateText"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginEnd="16dp"
+                android:gravity="center"
+                android:textColor="@color/md_theme_onBackground"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toStartOf="@id/exFourHeaderDivider"
+                app:layout_constraintStart_toStartOf="parent"
+                tools:text="Sab, 5 Desember" />
+
+            <View
+                android:id="@+id/exFourHeaderDivider"
+                android:layout_width="1dp"
+                android:layout_height="0dp"
+                android:background="@color/gray_spacer"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="@id/tv_start_from" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/text_sampai"
+                android:textColor="@color/md_theme_onBackground"
+                android:textSize="16sp"
+                android:layout_marginBottom="8dp"
+                app:layout_constraintStart_toStartOf="@id/exFourEndDateText"
+                app:layout_constraintEnd_toEndOf="@id/exFourEndDateText"
+                app:layout_constraintBottom_toTopOf="@id/exFourEndDateText"/>
+
+            <TextView
+                android:id="@+id/exFourEndDateText"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="16dp"
+                android:layout_marginTop="32dp"
+                android:gravity="center"
+                android:textColor="@color/md_theme_onBackground"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toEndOf="@id/exFourHeaderDivider"
+                app:layout_constraintTop_toBottomOf="@id/iv_close"
+                tools:text="Sen, 7 Desember" />
 
 
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+        <include
+            android:id="@+id/legendLayout"
+            layout="@layout/layout_calendar_day_legend_container"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="8dp"
+            android:layout_marginBottom="8dp" />
+
+    </com.google.android.material.appbar.AppBarLayout>
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <com.kizitonwose.calendar.view.CalendarView
+            android:id="@+id/exFourCalendar"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:clipToPadding="false"
+            android:paddingBottom="100dp"
+            app:cv_orientation="vertical"
+            app:cv_scrollPaged="false"
+            app:cv_dayViewResource="@layout/item_calendar_day"
+            app:cv_monthHeaderResource="@layout/item_calendar_header" />
+
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="76dp"
+            android:layout_gravity="bottom"
+            android:background="@color/md_theme_background"
+            android:elevation="10dp">
+
+            <Button
+                android:theme="@style/Example4ButtonTheme"
+                android:id="@+id/exFourSaveButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:layout_marginStart="16dp"
+                android:layout_marginEnd="16dp"
+                android:text="@string/text_simpan"
+                style="@style/Widget.Material3.Button"/>
+
+        </FrameLayout>
+
+    </FrameLayout>
+
+</LinearLayout>
