--- app/src/main/res/layout/example_4_fragment.xml
@@ -1,126 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/white"
-    android:clickable="true"
-    android:focusable="true"
-    android:orientation="vertical"
-    tools:context=".presentation.calenderfilterhistory.Example4Fragment">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/exFourAppBarLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/white"
-        android:theme="@style/AppTheme.AppBarOverlay">
-
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/exFourToolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            app:popupTheme="@style/AppTheme.PopupOverlay"
-            app:theme="@style/Example4ToolbarTheme"
-            app:titleTextColor="@color/md_theme_outline" />
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:paddingStart="14dp"
-            android:paddingTop="2dp"
-            android:paddingEnd="14dp"
-            android:paddingBottom="10dp">
-
-            <TextView
-                android:id="@+id/exFourStartDateText"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="30dp"
-                android:gravity="start|center_vertical"
-                android:minLines="2"
-                android:textColor="@color/md_theme_outline"
-                android:textSize="26sp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toStartOf="@id/exFourHeaderDivider"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                tools:text="Mon\n22 Apr" />
-
-            <View
-                android:id="@+id/exFourHeaderDivider"
-                android:layout_width="1.5dp"
-                android:layout_height="0dp"
-                android:background="#EAEAEA"
-                android:rotation="45"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-
-            <TextView
-                android:id="@+id/exFourEndDateText"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="30dp"
-                android:gravity="end|center_vertical"
-                android:minLines="2"
-                android:textColor="@color/md_theme_outline"
-                android:textSize="26sp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@id/exFourHeaderDivider"
-                app:layout_constraintTop_toTopOf="parent"
-                tools:text="Thu\n9 May" />
-
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <include
-            android:id="@+id/legendLayout"
-            layout="@layout/calendar_day_legend_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp" />
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <com.kizitonwose.calendar.view.CalendarView
-            android:id="@+id/exFourCalendar"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:clipToPadding="false"
-            android:paddingBottom="100dp"
-            app:cv_orientation="vertical"
-            app:cv_scrollPaged="false"
-            app:cv_dayViewResource="@layout/example_4_calendar_day"
-            app:cv_monthHeaderResource="@layout/example_4_calendar_header" />
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="76dp"
-            android:layout_gravity="bottom"
-            android:background="@color/white"
-            android:elevation="10dp">
-
-            <Button
-                android:theme="@style/Example4ButtonTheme"
-                android:id="@+id/exFourSaveButton"
-                android:layout_width="match_parent"
-                android:layout_height="60dp"
-                android:layout_gravity="center"
-                android:layout_marginStart="16dp"
-                android:layout_marginEnd="16dp"
-                android:text="@string/text_simpan" />
-
-        </FrameLayout>
-
-    </FrameLayout>
-
-</LinearLayout>
