+++ app/src/main/res/layout/item_calendar_day.xml
@@ -28,6 +28,7 @@
         android:layout_height="match_parent"
         android:gravity="center"
         android:textSize="16sp"
-        tools:text="22" />
+        tools:text="22"
+        android:textStyle="bold"/>
 </FrameLayout>
