+++ app/src/main/res/layout/item_calendar_day_legend_text.xml
