+++ app/src/main/res/layout/item_calendar_header.xml
@@ -9,6 +9,6 @@
     android:paddingTop="10dp"
     android:paddingEnd="16dp"
     android:paddingBottom="10dp"
-    android:textColor="@color/md_theme_outline"
     android:textSize="26sp"
+    android:textStyle="bold"
     tools:text="October 2019" />
