+++ app/src/main/res/values/strings.xml
@@ -88,4 +88,6 @@
     <string name="text_end_date">End Date</string>
     <string name="text_hapus">Hapus</string>
     <string name="text_calender_title_view">Calender View Filter</string>
+    <string name="text_mulai_dari">Mulai Dari</string>
+    <string name="text_sampai">Sampai</string>
 </resources>
